Add files via upload
This mini-project was completed as part of an internship application requirement. The goal was to use OpenCV to analyze a thermal video and identify components with a frequency of 0-1 Hz.
Project Overview
The project involved the following steps:

Acquisition of Thermal Video: Obtained a thermal video for analysis.
Preprocessing: Preprocessed the video to enhance the quality and prepare it for analysis.
Component Identification: Utilized OpenCV to identify connected components within the video.
Frequency Analysis: Analyzed the identified components to determine their frequency range.
Results
The project successfully identified components in the thermal video but every component had a frequency of less than 1 Hz. I think that is due to the result of random threshold values. I am unaware of how to choose those values. Kindly help me with that.